final 1

+++ Backend/Microservices-Backend/post-service/src/main/java/com/kharzixen/postservice/service/PostService.java
@@ -2,13 +2,15 @@ package com.kharzixen.postservice.service;
 
 import com.kharzixen.postservice.dto.ImageCreatedResponseDto;
 import com.kharzixen.postservice.dto.incomming.PostDtoIn;
-import com.kharzixen.postservice.dto.incomming.UserDtoIn;
 import com.kharzixen.postservice.dto.outgoing.PostDtoOut;
 import com.kharzixen.postservice.exception.PostNotFoundException;
+import com.kharzixen.postservice.exception.UnauthorizedRequestException;
 import com.kharzixen.postservice.exception.UserNotFoundException;
 import com.kharzixen.postservice.mapper.PostMapper;
 import com.kharzixen.postservice.model.Post;
+import com.kharzixen.postservice.model.PostOutbox;
 import com.kharzixen.postservice.model.User;
+import com.kharzixen.postservice.repository.PostOutboxRepository;
 import com.kharzixen.postservice.repository.PostRepository;
 import com.kharzixen.postservice.repository.UserRepository;
 import com.kharzixen.postservice.webclient.ImageServiceClient;
@@ -16,8 +18,10 @@ import jakarta.ws.rs.NotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.*;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
+import java.util.Objects;
 
 @Service
 @AllArgsConstructor
@@ -25,6 +29,7 @@ public class PostService {
     private final PostRepository postRepository;
     private final UserRepository userRepository;
     private final ImageServiceClient imageServiceClient;
+    private final PostOutboxRepository postOutboxRepository;
 
     public PostDtoOut getPostById(Long postId, Long requesterId) throws PostNotFoundException {
         Post post = postRepository.findById(postId).orElseThrow(() -> new PostNotFoundException(postId));
@@ -33,9 +38,9 @@ public class PostService {
         return postDtoOut;
     }
 
-    public PostDtoOut createPost(PostDtoIn postDtoIn) {
-        User owner = userRepository.findById(postDtoIn.getUploaderId())
-                .orElseThrow(() -> new UserNotFoundException(postDtoIn.getUploaderId()));
+    public PostDtoOut createPost(PostDtoIn postDtoIn, Long requesterId) {
+        User owner = userRepository.findById(requesterId)
+                .orElseThrow(() -> new UserNotFoundException(requesterId));
 
         ImageCreatedResponseDto imageResponseDto = imageServiceClient
                 .postImageToMediaService(postDtoIn.getImage(), postDtoIn.getUploaderId());
@@ -61,9 +66,20 @@ public class PostService {
                 posts.getTotalElements());
     }
 
-    public void deletePostById(Long postId) {
+    @Transactional
+    public void deletePostById(Long postId, Long requesterId) {
         // add outbox delete for media service
         Post post = postRepository.findById(postId).orElseThrow(() -> new PostNotFoundException(postId));
-        postRepository.delete(post);
+        if(Objects.equals(post.getOwner().getId(), requesterId)){
+            postRepository.delete(post);
+            PostOutbox postOutbox = PostOutbox.builder()
+                    .method("DELETE")
+                    .postId(post.getId())
+                    .imageId(post.getImageId())
+                    .build();
+            postOutboxRepository.save(postOutbox);
+        } else {
+            throw new UnauthorizedRequestException("Unauthorized");
+        }
     }
 }
